requirements.txt
Flask
easyocr
opencv-python
twilio
python-dotenv
Pillow
gunicorn
opencv-python-headless